Restructure tokens code
- Remove tokens/mappers.py, as it doesn't do much
- Move tokens views into views.py (use the same structure as we do
  for ubuntu.com)
- Move CLI commands to commands.py
