+++ webapp/app.py
@@ -1,23 +1,23 @@
+# Packages
 from canonicalwebteam.flask_base.app import FlaskBase
 
+# Local
+from webapp.views import (
+    index,
+    create_token,
+    delete_token,
+    get_tokens,
+    get_token,
+)
 from webapp.database import db_session
-from webapp.decorators import token_required
-from webapp.tokens.cli import token_cli
-from webapp.tokens.flask import tokens_blueprint
+from webapp.commands import token_group
 
 
 app = FlaskBase(__name__, "assets.ubuntu.com")
 
-# Register tokens blueprint and CLI
-app.register_blueprint(tokens_blueprint)
-app.cli.add_command(token_cli)
 
-
-@app.route("/")
-@token_required
-def index():
-    return {}, 200
-
+# Error pages
+# ===
 
 @app.errorhandler(400)
 @app.errorhandler(401)
@@ -34,7 +34,30 @@ def error_500(exception=None):
     return {"status": "500", "message": ""}, 500
 
 
+# Routes
+# ===
+
+app.add_url_rule("/", view_func=index)
+
+# Tokens
+app.add_url_rule("/tokens", view_func=get_tokens)
+app.add_url_rule("/tokens/<string:name>", view_func=get_token)
+app.add_url_rule(
+    "/tokens/<string:name>", view_func=create_token, methods=["POST"]
+)
+app.add_url_rule(
+    "/tokens/<string:name>", view_func=delete_token, methods=["DELETE"]
+)
+
+
+# Teardown
+# ===
 @app.teardown_appcontext
 def remove_db_session(response):
     db_session.remove()
     return response
+
+
+# CLI commands
+# ===
+app.cli.add_command(token_group)
