+++ webapp/auth.py
@@ -1,6 +1,10 @@
-from webapp.tokens.mapper import find_token
+from webapp.database import db_session
+from webapp.models import Token
 
 
 def authenticate(token):
     """Check if this authentication token is valid (i.e. exists)"""
-    return True if find_token(token) else False
+
+    return bool(
+        db_session.query(Token).filter(Token.token == token).one_or_none()
+    )
